+++ examples/risc-v/gcc_test/fft.h
@@ -1401,9 +1401,9 @@ fft_2pt_out_t fft_2pt_w_omega_lut(
 }
 
 typedef struct fft_iters_t{
-    uint32_t s;
-    uint32_t k;
-    uint32_t j;
+    uint16_t s;
+    uint16_t k;
+    uint16_t j;
 }fft_iters_t;
 #if NFFT==1024
 #define CEIL_LOG2_NFFT_PLUS_1 (LOG2_NFFT+1)
@@ -1412,12 +1412,12 @@ uint1_t s_last(fft_iters_t skj){
     return (skj.s == (CEIL_LOG2_NFFT_PLUS_1-1));
 }
 uint1_t k_last(fft_iters_t skj){
-    uint32_t m = (uint32_t)1 << skj.s;
+    uint16_t m = (uint16_t)1 << skj.s;
     return (skj.k == (NFFT-m));
 }
 uint1_t j_last(fft_iters_t skj){
-    uint32_t m = (uint32_t)1 << skj.s;
-    uint32_t m_1_2 = m >> 1;
+    uint16_t m = (uint16_t)1 << skj.s;
+    uint16_t m_1_2 = m >> 1;
     return (skj.j == (m_1_2-1));
 }
 #define FFT_ITERS_NULL_INIT {.s=1, .k=0, .j=0}
@@ -1433,7 +1433,7 @@ fft_iters_t next_iters(fft_iters_t skj){
                 skj.s += 1;
             }
         }else{
-            uint32_t m = (uint32_t)1 << skj.s;
+            uint16_t m = (uint16_t)1 << skj.s;
             skj.k += m;
         }
     }else{
